+++ src/shared/Cryptography/Sha1.cpp
