+++ src/shared/Cryptography/Sha1.h
@@ -15,35 +15,40 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
  */
 
-#ifndef _MD5_H
-#define _MD5_H
+#ifndef _AUTH_SHA1_H
+#define _AUTH_SHA1_H
 
 #include <cstdlib>
-#include <openssl/md5.h>
 #include "Common.hpp"
+#include <openssl/sha.h>
+#include "Cryptography/BigNumber.h"
 
-class MD5Hash
+class Sha1Hash
 {
-    public:
+public:
+    Sha1Hash();
+    ~Sha1Hash() = default;
 
-        MD5Hash();
-        ~MD5Hash() = default;
+    void UpdateFinalizeBigNumbers(BigNumber* bn0, ...);
+    void UpdateBigNumbers(BigNumber* bn0, ...);
 
-        void UpdateData(const uint8* dta, int len);
-        void UpdateData(const std::string & str);
+    void UpdateData(const uint8* dta, int len);
+    void UpdateData(const std::string & str);
 
-        void Initialize();
-        void Finalize();
+    void Initialize();
+    void Finalize();
 
-        uint8* GetDigest(void) { return mDigest; };
-        int GetLength(void) { return MD5_DIGEST_LENGTH; };
+    uint8* GetDigest(void) { return mDigest; };
+    int GetLength(void) { return SHA_DIGEST_LENGTH; };
 
-    private:
+    BigNumber GetBigNumber();
 
-        MD5_CTX mC;
-        uint8 mDigest[MD5_DIGEST_LENGTH];
+private:
+    SHA_CTX mC;
+    uint8 mDigest[SHA_DIGEST_LENGTH];
 };
 
-#endif      //_MD5_H
+#endif
