+++ src/shared/Cryptography/WowCrypt.cpp
